requirements.txt
# ── Web Framework ──────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4

# ── Database ────────────────────────────────────────────
pymongo==4.8.0

# ── Authentication ──────────────────────────────────────
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# ── AI / LLM ────────────────────────────────────────────
groq==0.9.0

# ── RAG / Embeddings ────────────────────────────────────
sentence-transformers==3.0.1
numpy==1.26.4

# ── PDF Processing ──────────────────────────────────────
PyPDF2==3.0.1

# ── Utilities ───────────────────────────────────────────
python-dotenv==1.0.1